simple_backend/src/task_tracker/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
